+++ CHANGELOG.md
@@ -1,19 +1,27 @@
 # Changelog
 
 
-# 0.1.4
+## 0.1.5
+
+- Made the HFCardCollectionViewCell and HFCardCollectionViewLayout open instead of public
+- Improved the HFCardCollectionViewDelegate
+- Package.swift added
+- Reordered the source files
+
+
+## 0.1.4
 
 - Bugfix for setting the background color or HFCardCollectionViewCell
 - Extends the HFCardCollectionViewDelegate
 - Removed the HFCardCollectionViewCellDelegate
 
 
-# 0.1.3
+## 0.1.3
 
 - HFCardCollectionViewDelegate now inherits from UICollectionViewDelegate
 
 
-# 0.1.2
+## 0.1.2
 
 - Renamed option bottomShouldScaleStackedCards to bottomStackedCardsShouldScale
 - Improving of the calculation for the moving cell